Added several libinput config options
In `weston.ini`, the new options are called `tap-and-drag`,
`tap-and-drag-lock`, `disable-touchpad-while-typing`,
`middle-emulation`, `left-handed`, `rotation`, `accel-profile`,
`accel-speed`, `scroll-method`, `natural-scroll`, and `scroll-button`. I
have successfully tested everything except for `rotation`, out of a lack
of hardware support.

weston now depends directly on libevdev for turning button name strings into
kernel input codes. This was needed for the `scroll-button` config
option. (weston already depends indirectly on libevdev through
libinput, so I figured people would be OK with this.) As a practical
matter for debian-style packagers, weston now has a build dependency on
libevdev-dev.

`enable_tap` has been deprecated in favour of `enable-tap`. Both work for
now. If `enable-tap` is present, `enable_tap` is ignored. An obnoxious
message is printed in the log if `enable_tap` is used.